requirements.txt
netaddr
Pillow
requests
dnspython
